+++ docs/_pages/releases.md
@@ -8,6 +8,10 @@ sidebar:
 ---
 ## Unreleased
 
+### What's new
+ 
+* Add `EqualApproximately` for collections of `INumber<T>` - [#413](https://github.com/AwesomeAssertions/AwesomeAssertions/pull/413)
+
 ### Fixes
 * Add missing `[NotNull]` attribute on the `Should()` methods for nullable value types - [#411](https://github.com/AwesomeAssertions/AwesomeAssertions/pull/411)
 * Add missing reason to some equivalency assertions - [#405](https://github.com/AwesomeAssertions/AwesomeAssertions/pull/405)